Added window title option to post title extension
