+++ ext/post_titles/config.php
@@ -5,4 +5,5 @@ abstract class PostTitlesConfig
 {
     public const VERSION = "ext_post_titles_version";
     public const DEFAULT_TO_FILENAME = "post_titles_default_to_filename";
+    public const SHOW_IN_WINDOW_TITLE = "post_titles_show_in_window_title";
 }
