+++ ext/post_titles/events/post_title_set_event.php
@@ -1,1 +1,13 @@
 <?php
+
+class PostTitleSetEvent extends Event
+{
+    public $image;
+    public $title;
+
+    public function __construct(Image $image, String $title)
+    {
+        $this->image = $image;
+        $this->title = $title;
+    }
+}
