+++ ext/post_titles/main.php
@@ -11,11 +11,17 @@ require_once "events/post_title_set_event.php";
 
 class PostTitles extends Extension
 {
+    public function get_priority(): int
+    {
+        return 60;
+    }
+
     public function onInitExt(InitExtEvent $event)
     {
         global $config, $database;
 
         $config->set_default_bool(PostTitlesConfig::DEFAULT_TO_FILENAME, false);
+        $config->set_default_bool(PostTitlesConfig::SHOW_IN_WINDOW_TITLE, false);
 
         if ($config->get_int(PostTitlesConfig::VERSION) < 1) {
             $this->install();
@@ -29,6 +35,15 @@ class PostTitles extends Extension
         if ($config->get_int(PostTitlesConfig::VERSION) < 1) {
             $database->Execute("ALTER TABLE images ADD COLUMN title varchar(255) NULL");
             $config->set_int(PostTitlesConfig::VERSION, 1);
+        }
+    }
+
+    public function onDisplayingImage(DisplayingImageEvent $event)
+    {
+        global $config;
+
+        if($config->get_bool(PostTitlesConfig::SHOW_IN_WINDOW_TITLE)) {
+            $event->set_title(self::get_title($event->get_image()));
         }
     }
 
@@ -59,10 +74,13 @@ class PostTitles extends Extension
         $sb = new SetupBlock("Post Titles");
         $sb->start_table();
         $sb->add_bool_option(PostTitlesConfig::DEFAULT_TO_FILENAME,"Default to filename", true);
+        $sb->add_bool_option(PostTitlesConfig::SHOW_IN_WINDOW_TITLE,"Show in window title", true);
         $sb->end_table();
 
         $event->panel->add_block($sb);
     }
+
+
 
     private function set_title(int $image_id, string $title)
     {
@@ -78,7 +96,11 @@ class PostTitles extends Extension
         $title = $image->title??"";
         if(empty($title) && $config->get_bool(PostTitlesConfig::DEFAULT_TO_FILENAME)) {
             $info = pathinfo($image->filename);
-            $title =  basename($image->filename,'.'.$info['extension']);
+            if(array_key_exists("extension",$info)) {
+                $title = basename($image->filename, '.' . $info['extension']);
+            } else {
+                $title = $image->filename;
+            }
         }
         return $title;
     }
