+++ ext/view/events/image_info_set_event.php
@@ -1,0 +1,12 @@
+<?php
+
+class ImageInfoSetEvent extends Event
+{
+    /** @var Image */
+    public $image;
+
+    public function __construct(Image $image)
+    {
+        $this->image = $image;
+    }
+}
