+++ ext/view/theme.php
@@ -20,8 +20,6 @@ class ViewImageTheme extends Themelet
     public function display_page(Image $image, $editor_parts)
     {
         global $page;
-
-        $page->set_title("Image {$image->id}: ".html_escape($image->get_tag_list()));
         $page->set_heading(html_escape($image->get_tag_list()));
         $page->add_block(new Block("Navigation", $this->build_navigation($image), "left", 0));
         $page->add_block(new Block(null, $this->build_info($image, $editor_parts), "main", 20));
