+++ themes/lite/view.theme.php
@@ -5,7 +5,6 @@ class CustomViewImageTheme extends ViewImageTheme
     public function display_page(Image $image, $editor_parts)
     {
         global $page;
-        $page->set_title("Image {$image->id}: ".html_escape($image->get_tag_list()));
         $page->set_heading(html_escape($image->get_tag_list()));
         $page->add_block(new Block("Navigation", $this->build_navigation($image), "left", 0));
         $page->add_block(new Block("Statistics", $this->build_stats($image), "left", 15));
